+++ plugins/gitops-profiles/src/components/ClusterTemplateCardList/ClusterTemplateCardList.tsx
@@ -15,7 +15,7 @@
  */
 
 import React from 'react';
-import { Grid } from '@material-ui/core';
+import Grid from '@material-ui/core/Grid';
 import ClusterTemplateCard from '../ClusterTemplateCard';
 
 interface Props {