+++ plugins/kubernetes-cluster/src/components/ApiResources/ApiResources.tsx
@@ -19,7 +19,7 @@ import { useEntity } from '@backstage/plugin-catalog-react';
 import { Table, TableColumn } from '@backstage/core-components';
 import { IAPIGroup } from '@kubernetes-models/apimachinery/apis/meta/v1';
 import { useKubernetesClusterError } from '../KubernetesClusterErrorContext/KubernetesClusterErrorContext';
-import { makeStyles } from '@material-ui/core';
+import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles(theme => ({
   empty: {