+++ plugins/newrelic-dashboard/src/Router.tsx
@@ -15,7 +15,7 @@
  */
 import { Entity } from '@backstage/catalog-model';
 import React from 'react';
-import { Button } from '@material-ui/core';
+import Button from '@material-ui/core/Button';
 import { NewRelicDashboard } from './components/NewRelicDashboard';
 import {
   useEntity,