+++ plugins/org/src/components/Cards/Meta/LinksGroup.tsx
@@ -16,12 +16,10 @@
 import { EntityLink } from '@backstage/catalog-model';
 import { IconComponent, useApp } from '@backstage/core-plugin-api';
 import LanguageIcon from '@material-ui/icons/Language';
-import {
-  ListItem,
-  ListItemIcon,
-  ListItemText,
-  Divider,
-} from '@material-ui/core';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemIcon from '@material-ui/core/ListItemIcon';
+import ListItemText from '@material-ui/core/ListItemText';
+import Divider from '@material-ui/core/Divider';
 import React from 'react';
 import { Link } from '@backstage/core-components';