+++ plugins/apache-airflow/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-apache-airflow
 
+## 0.2.22
+
+### Patch Changes
+
+- e8f026a: Use ESM exports of react-use library
+- Updated dependencies
+  - @backstage/core-components@0.14.2
+  - @backstage/core-plugin-api@1.9.1
+
 ## 0.2.21
 
 ### Patch Changes