+++ plugins/newrelic-dashboard/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-newrelic-dashboard
 
+## 0.3.7
+
+### Patch Changes
+
+- 1d81cdd: Added an optional ESLint rule - no-top-level-material-ui-4-imports -in new relic plugin which has an auto fix function to migrate the imports and used it to migrate the Material UI imports for plugins/newrelic-dashboard
+- e8f026a: Use ESM exports of react-use library
+- Updated dependencies
+  - @backstage/core-components@0.14.2
+  - @backstage/plugin-catalog-react@1.11.1
+  - @backstage/core-plugin-api@1.9.1
+  - @backstage/catalog-model@1.4.5
+  - @backstage/errors@1.2.4
+
 ## 0.3.6
 
 ### Patch Changes