+++ plugins/periskop/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-periskop
 
+## 0.1.30
+
+### Patch Changes
+
+- e8f026a: Use ESM exports of react-use library
+- Updated dependencies
+  - @backstage/core-components@0.14.2
+  - @backstage/plugin-catalog-react@1.11.1
+  - @backstage/core-plugin-api@1.9.1
+  - @backstage/catalog-model@1.4.5
+  - @backstage/errors@1.2.4
+
 ## 0.1.29
 
 ### Patch Changes