+++ plugins/rollbar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-rollbar
 
+## 0.4.32
+
+### Patch Changes
+
+- e8f026a: Use ESM exports of react-use library
+- Updated dependencies
+  - @backstage/core-components@0.14.2
+  - @backstage/plugin-catalog-react@1.11.1
+  - @backstage/core-plugin-api@1.9.1
+  - @backstage/catalog-model@1.4.5
+
 ## 0.4.31
 
 ### Patch Changes