+++ plugins/search-backend-node/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-search-backend-node
 
+## 1.2.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.21.5
+  - @backstage/backend-tasks@0.5.20
+  - @backstage/backend-plugin-api@0.6.15
+  - @backstage/config@1.2.0
+  - @backstage/errors@1.2.4
+  - @backstage/plugin-permission-common@0.7.13
+  - @backstage/plugin-search-common@1.2.11
+
 ## 1.2.18
 
 ### Patch Changes