+++ plugins/vault-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-vault-node
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.15
+
 ## 0.1.7
 
 ### Patch Changes